+++ lib/element.ts
@@ -7,6 +7,8 @@ import {
 } from '../util/string';
 import {typeAliases, classes} from '../util/constants';
 import {setChildNodes} from '../util/debug';
+import {isToken} from '../util/debug';
+import {generateForSelf} from '../util/lint';
 import {parseSelector} from '../parser/selector';
 import {Ranges} from './ranges';
 import {Title} from './title';
@@ -17,6 +19,7 @@ import type {
 	AstNodes,
 	AstText,
 	Token,
+	CategoryToken,
 } from '../internal';
 
 declare type TokenPredicate<T extends Token = Token> = (token: Token) => token is T;
@@ -287,9 +290,18 @@ export abstract class AstElement extends AstNode {
 	 */
 	lint(start = this.getAbsoluteIndex()): LintError[] {
 		const errors: LintError[] = [];
+		const cats = new Set<string>(),
+			isCategory = isToken<CategoryToken>('category');
 		for (let i = 0, cur = start + this.getAttribute('padding'); i < this.length; i++) {
 			const child = this.childNodes[i]!;
 			errors.push(...child.lint(cur));
+			if (isCategory(child)) {
+				if (cats.has(child.name)) {
+					errors.push(generateForSelf(child, {start: cur}, 'duplicated category'));
+				} else {
+					cats.add(child.name);
+				}
+			}
 			cur += String(child).length + this.getGaps(i);
 		}
 		return errors;