+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **Added**
 
 - Linting errors for [`LinkToken`](https://github.com/bhsd-harry/wikiparser-node/wiki/LinkToken) are now also reported from [`CategoryToken`](https://github.com/bhsd-harry/wikiparser-node/wiki/CategoryToken) and [`FileToken`](https://github.com/bhsd-harry/wikiparser-node/wiki/FileToken) where applicable
+- New category of linting errors: duplicated categories
 
 ## v1.3.9